rename to extension.ts instead of extension-common.ts
